작품삭제 alert

+++ semiProject/src/main/webapp/WEB-INF/views/user/uploadBookList.jsp
@@ -14,6 +14,7 @@
 <title>Insert title here</title>
     <link rel="stylesheet" href="/css/mypageDefault.css">
     <link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@48,400,1,0" />
+    <script src="/js/sweetalert.min.js"></script>
 </head>
 <style>
     .line-content{
@@ -127,7 +128,7 @@
 		                            <div class="book-date">작품 등록일 : <span class="book-date"><%=ub.getBookDate() %></span></div>
 		                            <div>
 		                                <a href="/updateBookFrm.do?bookNo=<%=ub.getBookNo() %>" class="btn bc6 book-btn">수정</a>
-		                                <button class="btn bc33 book-btn modal-open-btn" target="#test-modal" onclick="deleteUploadBook(this, <%=ub.getBookNo()%>, <%=ub.getBookWriter()%>);">작품 삭제</button>
+		                                <button class="btn bc33 book-btn modal-open-btn" onclick="deleteUploadBook(<%=ub.getBookNo()%>, '<%=ub.getBookWriter()%>');">작품 삭제</button>
 		                                <%--<a href="/deleteUploadBook.do?bookNo=<%=ub.getBookNo() %>&bookWriter=<%=ub.getBookWriter() %>" class="btn bc33 book-btn modal-open-btn">작품 삭제</a> --%>
 		                            </div>
 	                            </div>
@@ -141,7 +142,7 @@
         </div>
     </div>
     <%@include file="/WEB-INF/views/common/footer.jsp" %>
-    <div id="test-modal" class="modal-bg">
+<%--    <div id="test-modal" class="modal-bg">
       <div class="modal-wrap">
         <div class="modal-head">
           <h2>확인</h2>
@@ -154,13 +155,25 @@
           <a href="/deleteUploadBook.do" class="btn bc6 btn-pill">확인</a>
           <button class="btn bc33 modal-close btn-pill">취소</button>
         </div>
-      </div>
+      </div> --%>
       
       <script>
-			function deleteUploadBook(obj, bookNo, bookWriter) {
-				location.href="/deleteUploadBook.do?bookNo="+bookNo+"&bookWriter="+bookWriter;
-			}	
-      
+	      function deleteUploadBook(bookNo, bookWriter) {
+	          swal({
+	              title: '작품 삭제',
+	              text: "작품을 삭제하시겠습니까?",
+	              icon: 'warning',
+	              showCancelButton: true,
+	              /*confirmButtonColor: '#AACB73',
+	              cancelButtonColor: '#ccc',*/
+	              confirmButtonText: '확인',
+	              cancelButtonText: '취소'
+	          }).then(function(result) {
+	              if (result) {
+	                  location.href="/deleteUploadBook.do?bookNo="+bookNo+"&bookWriter="+bookWriter;
+	              }
+	          })
+	      }
       </script>
 </body>
 </html>
